# problem_5_resource/openapi.yml
openapi: 3.0.3
info:
  title: Employee & Task Management API
  description: |
    A comprehensive employee and task management system API built with Express.js and TypeScript.
    This API provides:
    - User authentication (JWT-based)
    - Employee management
    - Task management with assignment tracking
    
    All employee and task endpoints require JWT authentication.
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /tasks:
    get:
      summary: List user's tasks
      description: Retrieve tasks assigned to the current user with optional filters
      operationId: getTasks
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter tasks by status
          required: false
          schema:
            type: string
            enum: [todo, in_progress, done]
            example: "todo"
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of tasks per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
      responses:
        '200':
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Tasks retrieved successfully"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 50
                      page:
                        type: integer
                        example: 1
                      totalPages:
                        type: integer
                        example: 5
                      limit:
                        type: integer
                        example: 10
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    post:
      summary: Create a new task
      description: Create a new task (defaults to assigning to current user)
      operationId: createTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Task created successfully"
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /tasks/{id}:
    get:
      summary: Get task by ID
      description: Retrieve a specific task (must be owner or assignee)
      operationId: getTaskById
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Task retrieved successfully"
                  data:
                    $ref: '#/components/schemas/Task'
        '403':
          description: Forbidden - not owner or assignee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    put:
      summary: Update task
      description: Update an existing task (must be owner or assignee)
      operationId: updateTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Task updated successfully"
                  data:
                    $ref: '#/components/schemas/Task'
        '403':
          description: Forbidden - not owner or assignee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    delete:
      summary: Delete task
      description: Delete an existing task (must be owner or assignee)
      operationId: deleteTask
      tags:
        - Tasks
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          description: Forbidden - not owner or assignee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  responses:
    UnauthorizedError:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Missing or invalid Authorization header"
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Task title
          example: "Complete onboarding"
        description:
          type: string
          description: Detailed task description
          example: "Read company handbook and complete setup"
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, done]
          description: Current status of the task
          example: "todo"
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority level
          example: "high"
        due_date:
          type: string
          format: date
          description: Task due date (YYYY-MM-DD)
          example: "2026-01-21"
          nullable: true
        assigned_to_user_id:
          type: integer
          description: ID of user assigned to this task
          example: 2
        created_by_user_id:
          type: integer
          description: ID of user who created this task
          example: 1
        created_at:
          type: string
          format: date-time
          description: Timestamp when task was created
          example: "2026-01-14T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when task was last updated
          example: "2026-01-14T10:30:00Z"
      required:
        - id
        - title
        - status
        - priority
        - assigned_to_user_id
        - created_by_user_id
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Task title
          example: "Complete onboarding"
          minLength: 1
        description:
          type: string
          description: Detailed task description
          example: "Read company handbook and complete setup"
        status:
          type: string
          enum: [todo, in_progress, done]
          description: Initial status (defaults to 'todo')
          default: todo
          example: "todo"
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority (defaults to 'medium')
          default: medium
          example: "high"
        due_date:
          type: string
          format: date
          description: Task due date (YYYY-MM-DD)
          example: "2026-01-21"
        assigned_to_user_id:
          type: integer
          description: ID of user to assign (defaults to current user)
          example: 2
      required:
        - title
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Task title
          example: "Complete onboarding process"
          minLength: 1
        description:
          type: string
          description: Detailed task description
          example: "Updated description"
        status:
          type: string
          enum: [todo, in_progress, done]
          description: Task status
          example: "in_progress"
        priority:
          type: string
          enum: [low, medium, high]
          description: Task priority
          example: "high"
        due_date:
          type: string
          format: date
          description: Task due date (YYYY-MM-DD)
          example: "2026-01-21"
        assigned_to_user_id:
          type: integer
          description: ID of user to assign
          example: 3
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
...